Changed file layout
