+++ 2048Game/Number.cpp
@@ -1,18 +1,4 @@
-
-class Number
-{
-private:
-     int number;
-     bool hasJoinedInThisTurn;
-public:
-    int GetNumber();
-    void SetNumber(int value);
-    void Join(Number number);
-    bool Compare(Number number);
-    void ResetTurn();
-    Number();
-    ~Number();
-};
+#include "Number.h"
 
 Number::Number()
 {
